created tcp module and moved tcp connection setup into it
